fix: optimize dashboard queries to handle 3M+ schedules and fix InvalidCastException
Fixed critical dashboard performance issues causing infinite spinner and InvalidCastException errors with 3M+ schedules.

**Problems:**
1. InvalidCastException: JobExecution entity had 'new long Id' but database has INT
2. Dashboard Performance: Every query loaded ALL 3M+ schedules into memory, then filtered
3. Dashboard Timeout: Queries took too long, causing UI to timeout with spinner

**Solutions:**

1. **Fixed InvalidCastException:**
   - Removed 'new long Id' from JobExecution entity (line 9)
   - Changed JobExecutionRepository.GetByIdAsync(long id) -> GetByIdAsync(int id)
   - Now correctly maps to INT column in database

2. **Optimized Dashboard Queries:**
   - **GetOverview**: Changed from GetAllAsync() to FindAsync() with filters
   - **GetStatusBreakdown**: Use server-side JOIN on Schedule.ClientId instead of loading all schedules
   - **GetExecutionTrends**: Use server-side JOIN on Schedule.ClientId instead of loading all schedules
   - **GetTopLongestExecutions**: Use server-side JOIN and Take(limit) before ToList()
   - **GetInvalidSchedules**: Added limit parameter (default: 100) to prevent scanning all 3M+ schedules

3. **Key Optimization Pattern:**
   - Before: Load all 3M+ schedules -> filter in memory -> intersect with executions
   - After: Filter executions with WHERE e.Schedule.ClientId = @clientId (server-side JOIN)
   - Result: EF generates efficient SQL JOIN instead of loading everything

4. **Database Indexes:**
   - Added IX_JobExecutions_StartTime_Status_Duration (covering index)
   - Added IX_JobExecutions_ScheduleId_StartTime
   - Supports dashboard queries with StartTime, Status, ScheduleId filters

**Performance Impact:**
- Before: Dashboard never loaded (infinite spinner, timeout after minutes)
- After: Dashboard should load in seconds with proper indexes
- Memory: Reduced from loading 3M+ schedules to only loading filtered executions

**Files Changed:**
- JobExecution.cs: Removed 'new long Id' override
- JobExecutionRepository.cs: Changed GetByIdAsync signature to int
- DashboardController.cs: Rewrote all queries to use server-side filtering
- Migration: Added dashboard performance indexes
- SQL script: Manual index creation for immediate deployment

**Testing Required:**
- Verify dashboard loads successfully
- Check query execution times in SQL profiler
- Confirm indexes are being used (check execution plans)
- Monitor memory usage during dashboard load

**Deployment:**
- Run migration: dotnet ef database update
- OR run add_dashboard_indexes.sql manually
- Restart API to pick up entity changes

+++ src/SchedulerPlatform.Infrastructure/Repositories/JobExecutionRepository.cs
@@ -12,7 +12,7 @@ public class JobExecutionRepository : Repository<JobExecution>, IJobExecutionRep
     {
     }
 
-    public async Task<JobExecution?> GetByIdAsync(long id)
+    public async Task<JobExecution?> GetByIdAsync(int id)
     {
         var execution = await _dbSet
             .Include(je => je.Schedule)
